Add files via upload
The program was built for a very specific type of data format. It reads an excel file with the load in the second column and displacement in the fourth. It also requires the user input of weight and a breaking point. The program then calculates the maximum load held, the strength (load over bridge weight), and the tension (the product of load and displacement). Since the load is not constant, the program uses basic Reiman sums to calculate the integral under the curve of a load versus displacement graph manually. The program then prints this information to the console and to a new save file for the user.